Remove `fork` in favor of passing options to `spawn`
